Renames component scripts to have .jsx extensions
