+++ src/Components/Common/index.js
@@ -1,4 +1,6 @@
-import { List } from './List';
-import { ListElement } from './ListElement';
+import List from './List';
+import ListElement from './ListElement';
+import Popper from './Popper';
+import Portal from './Portal';
 
-export { List, ListElement };
+export { List, ListElement, Popper, Portal };
