+++ tests/storybook/components/label.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs } from '@storybook/addon-knobs';
-import { Label, LabelInput } from 'Components/Forms';
-import { LabelSet } from 'Components/Forms';
+import { Label, LabelInput, LabelSet } from 'Components/Forms';
 
 storiesOf('Label', module)
   .addDecorator(withKnobs)
